requirements.txt
geopandas==0.9.0
matplotlib==3.5.1
numpy==1.21.5
pandas==1.4.1
python_dateutil==2.8.2
scikit_learn==1.1.2
seaborn==0.11.2
Shapely==1.7.1
statsmodels==0.13.2
